tweetquizzes/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.6b0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.4
gunicorn==20.1.0
idna==2.10
mecab-python3==1.0.3
mypy-extensions==0.4.3
oauthlib==3.1.1
pathspec==0.8.1
py==1.10.0
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.5
django-heroku
unidic-lite
